# ec_backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/download": {
      "get": {
        "operationId": "S3Controller_getObject",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The object has been successfully fetched."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "S3"
        ]
      }
    },
    "/api/upload": {
      "post": {
        "operationId": "S3Controller_putObject",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The object has been successfully uploaded."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "S3"
        ]
      }
    },
    "/api/delete": {
      "delete": {
        "operationId": "S3Controller_deleteObject",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The object has been successfully deleted."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "S3"
        ]
      }
    },
...
    "/api/slideshow/slides": {
      "post": {
        "operationId": "SlideController_generateSlide",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSlideDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "example": "The slide has been created!"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "The slideshowId field must be a UUUID."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Slide"
        ]
      }
    },
    "/api/slideshow/slides/{slideId}": {
      "patch": {
        "operationId": "SlideController_updateSlide",
        "parameters": [
          {
            "name": "slideId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSlideDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The slide has been updated."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Validation failed (uuid is expected)"
              }
            }
          },
          "404": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "The ID is invalid or the slide does not exist."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Slide"
        ]
      },
      "delete": {
        "operationId": "SlideController_deleteSlide",
        "parameters": [
          {
            "name": "slideId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "The slide has been deleted."
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "Validation failed (uuid is expected)"
              }
            }
          },
          "404": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": "The ID is invalid or the slide does not exist."
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": "An unexpected error occurred. Please check the error log for more information."
              }
            }
          }
        },
        "tags": [
          "Slide"
        ]
      }
    },
...
  },
  "info": {
    "title": "English Conversations API",
    "description": "REST API for managing English Conversations' data and processes.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "GenerateSlideDTO": {
        "type": "object",
        "properties": {
          "slideshowId": {
            "type": "string",
            "title": "SlideshowID",
            "example": "0df0a1e2-1347-4d19-8d93-7aa0c29f9329"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "example": "Unit 1 - Past Simple"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "example": "Learn the past simple tense easily."
          },
          "type": {
            "type": "string",
            "title": "Type",
            "example": "IMAGE"
          },
          "url": {
            "type": "string",
            "title": "URL",
            "example": "https://google.com/images"
          },
          "order": {
            "type": "number",
            "title": "Order",
            "example": 1
          }
        },
        "required": [
          "slideshowId",
          "title",
          "description",
          "type",
          "url",
          "order"
        ]
      },
      "UpdateSlideDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "example": "Unit 1 - Past Simple"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "example": "Learn the past simple tense easily."
          },
          "type": {
            "type": "string",
            "title": "Type",
            "example": "IMAGE"
          },
          "url": {
            "type": "string",
            "title": "URL",
            "example": "https://google.com/images"
          },
          "order": {
            "type": "number",
            "title": "Order",
            "example": 1
          }
        },
        "required": [
          "title",
          "description",
          "type",
          "url",
          "order"
        ]
      },
...
    }
  }
}